optimize code.

+++ common/ui-dialog/src/main/java/com/app/base/ui/dialog/dsl/Indicator.kt
@@ -2,22 +2,22 @@ package com.app.base.ui.dialog.dsl
 
 class Indicator {
 
-    private var _offsetFromEnd: Int = -1
+    private var _offsetFromEnd: Float = -1F
 
-    private var _offsetFromStart: Int = -1
+    private var _offsetFromStart: Float = -1F
 
     @Direction private var _direction: Int = Direction.TOP
 
     /** -1 means indicator is at center. */
-    fun offsetFromEnd(offset: Int) {
-        _offsetFromStart = -1
+    fun offsetFromEnd(offset: Float) {
+        _offsetFromStart = -1F
         _offsetFromEnd = offset
     }
 
     /** -1 means indicator is at center. */
-    fun offsetFromStart(offset: Int) {
+    fun offsetFromStart(offset: Float) {
         _offsetFromStart = offset
-        _offsetFromEnd = -1
+        _offsetFromEnd = -1F
     }
 
     fun direction(@Direction direction: Int) {
@@ -35,7 +35,7 @@ class Indicator {
 }
 
 class IndicatorDescription(
-    val offsetFromEnd: Int,
-    val offsetFromStart: Int,
+    val offsetFromEnd: Float,
+    val offsetFromStart: Float,
     @Direction val direction: Int,
 )
